Move actor-related patches into a separate file
