+++ MWSE/PatchActors.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace mwse::patch::actors {
+	void install();
+}
